2.2.0 - Support YAML format for card list.
Switch from JSON to YAML since it doesn't require quotes around strings
and therefore is less confusing for cards with quotes in their names.
It's also generally more compact and easier to read. This is optional,
though. A new advanced setting was added to allow switching back to
JSON. The whole watchlist can still be edited directly in the script
manager's values section, and is saved there in JSON format. Cards now
include a rarity field. The script doesn't use it for anything but it
may help for your own organization. It also supports a notes field,
which can be used to store any extra information you want. It won't be
read aloud. There's also now an alert field, which is similar, but is
read aloud when the card is found, along with its name. So if you want
to add something to get your attention for a specific card, you can use
the alert field for that. These values are optional and won't be added
automatically when cards are selected by clicking. Also fixed some
issues with handling certain unusual pages like Boosters and Redeemable
Sets, as well as how card names are spoken on those pages. The Escape
key now closes all the script's dialogs. Finally, improve the code
comments documenting the script's settings.